Add metainfo.xml files for pst-import plugin
This is useful for GNOME Software, to show the plugin as an addon
for the Evolution. Packagers can decide what to do with the file,
either they can add it into the respective subpackage or they can
just ignore (remove) the file, if they do not build special package
for this plugin.