EHTMLEditorView - Redoing the Return key press in the list can fail
This could happen when the Return key is pressed in the last item of the list
and that item is also empty (usually this is done to end the list). When redoing
this event the DOM structure is restored wrongly. To correct this bug we have to
mark the history event correctly.