EHTMLEditorView - Pressing the Return key in the bulleted list does not end it
The Return key pressed in an empty item on the end of the bulleted list does not
end the list (as it is done by WebKit for other lists). This is caused by CSS
modifications of the bulleted list in the plain text mode (to have the "*"
character instead of bullet).