Bug 761394 - Composer does not remember if a signature was set to 'None' in Drafts/Templates
Reflect in the HTML code of a message that there is no signature set, so we
behave consistently when opening a message from drats or templates. Improve the
code to create the DOM structure of the signature instead of constructing it
from the string. Also save the signature ID to the element id attribute instead
of the name attribute (as element's id was set to '1' but not never used).