[EMinicard] Release grab when started drag of the contact
As discussed in bug #760951, drag & drop from the minicard view
into the source selector results in a state where there cannot be
clicked anything, but the minicard view again. That was due to left
grab on the minicard view canvas, while it should be released,
the same as it is released on button release in other cases.