Bug 761496 - Unicode emoticon requires two backspace presses to delete
This does not apply just on the Unicode emoticons, but for the image ones as well.
While working on this I reworked hot the Unicode smileys are inserted to the
composer. Before they were inserted inside a SPAN element and this caused some
problems, after this change they are inserted as text.