Bug 762376 - Embedded images in signature only displays under evolution and not other clients
The main problem is that the GMail does not support base64 images. This is not a
problem for Evolution as it is using them just for the Drafts and Templates. But
there was a problem with the signatures as when they were saved the images inside
them were left in the base64 format, but the data-inline attribute was removed
from them. This later prevented the composer to correctly handle them.

To avoid this issue change how the signatures are saved in the Signature editor.
Save them in the same way as the Drafts and don't try to process them when saving
them, but when the message with the given signature is being send. This could also
fix issues with editing a signature once it is in the body.