Correct packing of the EHTMLEditorView in the EMsgComposer
The WebKitWebView is not a scrollable anymore, thus the GtkScrollableWindow
packs it into a GtkViewport. That means the direct parent of the view is
the viewport, not the scrollable window. Count with this in the code and
show a runtime warning if any of these assumptions fail in the future.