Plain text signature is opened as an HTML one
After the commit cca2aa2 signatures are always saved in HTML. Remember the
editor mode to be able to switch to the right one after editing the signature
again.

Also force the font in the preview to be monospaced for the signatures that were
saved in the plain text mode.