Move the clipboard handling from EMsgComposer to EHTMLEditorView
Before this change the clipboard handling would be different when the view was
embedded inside a composer or e.g. in the signature editor. Move the handling
to EHTMLEditorView to unify the behavior for most cases apart from pasting an
image or uri in the plain text mode as the content is inserted to the attachment
view.