Recognize special folders for a global mail view also based on flags
Folders which are reported as Sent or Outbox with their folder info
flags only were not recognized for proper global folder view. That
made for example show a From column, instead of the To column, where
the former is rather useless in the Sent or Outbox folders.

This had been reported at:
https://bugzilla.redhat.com/show_bug.cgi?id=1332754