EHTMLEditorView - Simplify and improve how the undo/redo of delete operation in quoted content is performed
After processing the first and the last block in the deleted content insert the
fragment with the rest as a whole to the given position and let the
merge_siblings_if_necessary to do its work. Previous behaviour could skip some
blocks or insert them into different position or reverse them. With this
approach there is no need for marking the blocks with the special id and treat
them separately.