EHTMLEditorView - Deleting a content in a PRE element could wrap the content in SPAN element
When deleting a content from a PRE element WebKit could wrap the content after
the caret and the end of the element into the SPAN element when a special style
attribute is set on the BODY. This could lead to various breaks in the composer.
To avoid it we need to rename the style attribute when deleting a content in the
plain text mode.