Finish the second leftover (of 2) from the EAttachmentBar changes (Do not access EAttachmentStore from EAttachment)
Storing the GtkTreeReference on the attachment works only if
the attachment is inside a single store. When it is in multiple
stores, then the changes done on the attachment "view-related"
properties are propagated to all windows, which is not as it
should be. Also the GtkTreeReference is overwritten with the latest
store, thus it breaks other things too.