Rely more on CSS when modifying elements for composer mode change
Introduce an data-evo-plain-text BODY attribute that is presented only when
the composer is in the plain text mode. Based on this attribute we've modified
the CSS rules that let us in the end to be able to remove some unneeded CSS
classes and also some functions like toggle_unordered_lists.