Bug 750299 - [regression] Hyperlinks create duplicate text, then deleting the duplicate makes the hyperlink blow up to the whole line
Fix the EHTMLEditorLinkDialog as between showing the dialog and clicking the "OK"
button the selection in composer was lost that later lead to creating the duplicate
text. Also save the link under the cursor straight in the EHTMLEditor and then pass
it to the dialog to avoid looking for it on various places again.

Also it showed up that I completely forgot about this dialog when implementing the
undo/redo in the composer, so previously doing so after working with dialog would
lead to undefined behavior.

Replace the webkit_dom_html_anchor_element_get_href with webkit_dom_element_get_attribute
as the first one would return just "http:/" when we would previously set the "http://"
as a value of the href attribute.