Introduce a new test for pressing a Return key after the HR element
And fix the code to do it correctly - previously no block was inserted after the
HR element thus the user could write straight into the BODY element (but this is
harmless as there is a code that wraps the text into the block). Also fix the
undo and redo for this action.