Add TestKeyfileSettingsBackend for test-html-editor-units
It used a 'memory' GSettings backend, but this one is not shared
between processes, thus the UI process settings changes were not
propagated into the WebProcess and vice versa, which could break
tests easily (notably the /emoticon tests, which expect magic-
smileys being set, but the default is unset). This test keyfile
settings backend shared settings through a temporary file between
the processes, as expected.