Ensure EMailConfigPage::changed signal is emitted in the main/UI thread
The signal is emitted also on any CamelSettings changes of the associated
CamelService, but some providers, like evolution-mapi, can change the profile
name of the settings in a dedicated thread. That could lead to a crash, due to
the EMailConfigAssistant listening to this signal and doing gtk+ calls there.

This had been reported downstream at:
https://bugzilla.redhat.com/show_bug.cgi?id=1368612