Some editor unit tests fail after commit d3fc7115732b25e9f3c1800c1a6b4902824c1cd0
Some of the EHTMLEditor actions check whether the content editor is focused,
to prevent running the action when for example the Subject entry is focused,
which broke unit tests with the window being in the background, because
the content editor widget didn't have set the 'focused' flag, even the editor
window has the content editor widget set as being focused. The new code reflects
both variants.