Undoing a citation split could fail
When the caret is on the end of the citation and the Return key is pressed then
a new empty block is inserted after the current block in citation. When we later
try to undo this operation the newly added block will be removed as it should,
but the next block would be concatenated after the current citation.