Use DIV instead of P element for 'Normal' block
While working on the WK2 port I decided to use P element for 'Normal' block
instead of DIV element. Now, it showed up that it was not a wise change as the
P element usually has some style defined by the web engine itself (usually
margins above and below) so the HTML message written in the Evolution's composer
won't look the same in other MUAs and actually not in the Evolution's preview as
well. Instead of adding some style attributes to every 'Normal' paragraph that
is being send basically revert the change and return back to DIV elements.