Bug 771420 - [WK2] Printing an email does not work
Adapt the EMailPrinter to the WebKit2 API. While doing so port it from using
GSimpleAsyncResult to GTask. Also we need to make some changes to EWebView as
if the print dialog was showed and immediately closed several times in row then
it could happen that the callbacks from EWebView could be called without a valid
EWebView object. This was addressed by using the GWeakRef.