Bug 772171 - Editing quoted replies can break message content
Don't save the selection again in
e_editor_dom_fix_structure_after_delete_before_quoted_content() as later
the selection markers are not used to restore the selection, but they
would prevent the delete operation to process correctly.

Also restore the selection on other places when we are not using it
anymore.