Disable/remove WebKit1 left artifacts of ENABLE_CONTACT_MAPS
The code which showed contact map in the contact preview required
WebKit1 functionality (gtk+ widgets inside webview), but it's gone
in WebKit2. This didn't break in the compile time, but in runtime
a critical warnings about missing signal had been shown on the console.
Let's disable/remove the related code for now and enable contact
maps in the preview once there's available a good replacement.
All the other functionality still works, thus the overall configure
option is left there.