Use unique D-Bus service name for WebKit editor and cancel pending calls on dispose
The former part is good to not reuse D-Bus service name between different processes.
The later part fixes a crash of `test-html-editor-units -p /font/size` where
the test could crashed right after the first finished. The reason was that
the dispose call on the window caused style change signal, which called
ElementSetAttributeBySelector D-Bus method, which may not always finish before
the the WebKitWebProcess stopped, which resulted in a D-Bus error and following
test abort due to a critical warning with that particular error message.