Process the WebView's context menu actions synchronously
This will prevent the WebView's context menu to be sized properly on
Wayland (on X11 the code is working fine, but it is not correct). There
is still one case where have to do an async call (inside the
EFocusTracker to get the clipboard's content) that will leave us with an
empty item in the context menu is some cases. This one will be fixed by
upstream gtk+ bug [0].

This was originally reported downstream [1].

[0] - https://bugzilla.gnome.org/show_bug.cgi?id=772505
[1] - https://bugzilla.redhat.com/show_bug.cgi?id=1394993