EEditorDOMFunctions: Critical warning and possible crash when we cannot append the end node for spell checking
If we fail to obtain a parent node of the last element in the viewport
we still tried to append to it, but that only caused a critical warning,
but as a consequence a WebKitDOMRange object was not created and we
tried to unref it that lead to a crash. Be sure that we always append
the spell check end node.