Bug 775397 - Freeze / lockup trying to delete a hyperlink in a cited html mail
Switch to the safer check whether we are after the ending range when
refreshing the spell check - compare the ranges start positions and
quit the loop after we move after the start position of the ending
range. Previously we needed to get to the exact ending position and that
was sometimes not possible (it could happen that it was skipped - it
depends on the DOM structure) and we continued to move the selection
over and over.