Bug 774494 - ENameSelectorDialog slow with large address books
Wait until contacts are fetched before attaching a GtkTreeModelSort.
Similarly, when changing the view during a search, detach the
GtkTreeModelSort while updating the tree view, and use a hash table
to temporarily store contacts, as this speeds up searching.

Also, in ETreeModelStore, add a cache to optimize
generated_offset_to_child_offset to avoid needing to iterate through the
whole list to count rows when a value is needed.