Possible deadlock in the replace dialog
We could get into the state where the Evolution was waiting until some
web process operation will finish, but the web process was waiting for
something from the UI Process (Evolution) - in this case spell check.
To avoid it move to the next word while doing 'replace all' only on
idle.