[Mail] Cached remote content not always found in cache
The URL being used to check whether the image is already locally cached
and the URL which was used to actually download the remote content did not
match, thus the file were not found in the local cache. While I've been in
it I also noticed that the remote content is not shown in the composer,
even when it's downloaded locally, thus I fixed that as well.

Reported downstream at:
https://bugzilla.redhat.com/show_bug.cgi?id=1435288