Drag and drop to the editor can lock the Evolution
The problem was that we were not disconnecting the drag-data-received
handler in the composer if the drop was handled by editor. Then in the
next drag and drop operation the callbacks in the composer were
activated even before the drag and drop logic in the editor. Introduce
the drop-handled signal that is emitted on the EContentEditor if the
drop was successfully handled by it. In the composer we connect to this
signal and remove the callback to the drag-data-received signal.