Dropping text/* content to the editor is not handled by WebKit
The problem was that the WebKit failed to process the drag-drop signal as
it internally failed to found its corresponding object to the
GdkDragContext that was passed to it. The problem was that the
corresponding object was removed when the drag-leave signal was
processed by WebKit as it internally tries to postpone the drag-leave
signal after the drag-drop is processed. As a solution we need to
connect to the drag-leave signal, stop its emission and emit at after
the drag-drop.

In case where the drop is not handled by WebKit warn to the console (as
this is not supposed to happen) and try to handle the drop ourselves.