[ENameSelectorEntry] Avoid fake "changed" signals on focus-in
When the name selector entry is focused, it changes its content by
appending ", " to make it easier to add new contacts, but it also
emits the "changed" signal, to which listens EMsgComposer. It's fine
as long as this doesn't happen when user wants to close the composer
while staying in the name selector entry (To/Cc/Bcc) with some address
filled. The thing is that the fake "changed" signal marks composer
dirty again, which avoids its close after draft save.