Fix occasional runtime warning from e_web_extension_set_clipboard_flags()
The WebKitDOMDocument had not set page ID, because it had not been
loaded yet, thus the runtime check in e_web_extension_set_clipboard_flags()
failed, because the callback to "selection-changed" is registered
early. Either stop to claim and treat no page ID as a valid case,
or set page ID in web_page_created_cb() too. This does the later.