Bug 780088 - Pasted URL not recognized/rendered as such; renders all remaining email text as link after pressing Enter
The main cause for both problems mentioned in the bug was the check for
the non-breaking space in the regex that we are using to recognizing the
URLs. Making it work would need a support for regexes in the negative
lookbehind, but GRegex does not support them. In the end I decided to
remove the check and check for the non-breaking space character in the
code rather than in the regex itself.