MailFolderCache can sometimes block folder 'changed' notifications
One way to have lost folder 'changed' notifications from MailFolderCache
is to fail to connect to the account after start, in which case the folder
tree in the Mail view doesn't update unread counts properly. It had been
due to the MailFolderCache still considering the store being initializing
for the first time, while piling the folder notifications to be emitted
when the initialization of the store is over.