Make it possible to change theme for text-highlight module
Since now, users can change theme being used by the text-highlight module,
not being forced to use a hard-coded theme in the code. First of all,
figure out which themes are installed. It's done with command:

   $ highlight --list-scripts=themes

Pick one, say "editor-vim-dark" (supposing it's installed), and execute:

   $ gsettings set org.gnome.evolution.text-highlight theme edit-vim-dark

And the next time the text-highlight will be called, like when changing the used
format with right-click in the body->Format As->..., the changed theme will
be used. In case the value being set in the GSettings is no longer valid,
like when the theme is not available or had been uninstalled, there will be
no formatting in the message content at all and evolution's console may
contain an error message from highlight, like this one:

   highlight: cannot open themes/unknown.theme: No such file or directory