Derive date from selected day when creating event/meeting in To Do bar
Right-click and create new event or task already derives the calendar/task
list of the selected component. This derives also selected date, thus for
example creating a new event and right-clicking on the 'Tomorrow' row
the event will be prefilled with the Tomorrow date and the nearest
future time of the current time according to set time division for
the Day and Work Week views.