Workaround broken GPG encrypted messages from Exchange and GroupWise servers
An Exchange server, eventually also a GroupWise server, can break a GPG
encrypted message structure by changing it from

   multipart/encrypted; protocol="application/pgp-encrypted"
      application/pgp-encrypted
      application/octet-stream; name="encrypted.asc"

to something like:

   multipart/mixed
      text/plain
      application/pgp-encrypted; name="ATT00001"
      application/octet-stream; name="encrypted.asc"

where the text/plain part is empty and is added by the Exchange server.
The added code tries to detect such issues and if recognized, then
changes the structure to be able to decrypt the message as expected.