Custom page setup not properly restored when printing
Using standard and then custom page size saved it into the printing.ini
with keys related to standard page size and the print dialog preferred
the standard size over the custom. Removing the group from the key file
first makes this work properly. Do the same for printer settings.
Also allow changing page size in the print dialog, but do not save it.