Remember back up file extension and default to .tar.gz again
The default had been .tar.xz, but as claimed in [1], it makes back up
significantly slower with not much gain on the back up file size. This
change adds a setting to remember what file type had been used the last
time and it is reused the next time the back up is invoked.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1491024