Label shortcuts prevent typing numbers
Since change for bug 787002 typing numbers into for example search
entry was not possible, due to the shortcuts for labels "consuming"
the key presses. The fix is to allow labels only if the message
list is focused, when also the context menu contains the Labels.

This had been reported downstream:
https://bugzilla.redhat.com/show_bug.cgi?id=1498484