Revert UI part of the change for bug 788916 (Use of localized Fwd/Re)
Make the change to use localized Fwd/Re hidden from the UI, because
it can break things in an international environment and the RFC 5256
section 7 suggests not to localize those. As this is eventually bad
option, it should not be shown in the UI (there are more interesting
options hidden in gsettings without their UI counter parts too).