[Composer] Better care of autosave snapshot after WebKitWebProcess crash
This is related to bug 791291, but doesn't fix it, it rather makes sure that:
a) any autosave snapshot is not overwritten with empty body of the message
b) the snapshot is not deleted after "broken" composer close
c) it's offered to recover the last snapshot (if exists) on close of the composer.

This all makes sure that users will not lose all their work, only up
to one minute of time, which is the interval of the autosave.