Drag&drop of attachment button from mail preview does nothing
It's using the same mime type, but the code to make it a real file
or anything like that fails. With this change the button (or the image
of it on the right side of the button) can be dragged into the folder
tree or message list, importing it into the mail folder.