Hide notification when other clients read mail
If a user checks new messages on another device, they don't need a
notification on this one reminding them about it. So this commit adds an
event emitted when the number of unread messages in a mail folder
changes. The mail-notification plugin uses that event to remove the
notification if the number of unread messages in a folder ever
decreases, implying that the user has checked their mail (or at least
has interacted with their mail server somehow) elsewhere.