[EMailDisplay] Do not call reload when nothing is loaded
It can happen that the EMailDisplay's reload, due to property changes,
is scheduled for idle during the EMailBrowser construction and that
idle callback is called when WebKitGTK+ doesn't have loaded expected
URI in it, only "about:blank". Calling reload in such state can prevent
the real (and possibly ongoing) request to be cancelled, which results
in an empty message preview. Avoiding schedule of such false reload
helps to prevent this situation.

This had been reported downstream at:
https://bugzilla.redhat.com/show_bug.cgi?id=1560312