[composer-autosave] Use-after-free during snapshot save to file ][
Extend the previous change to not do autosave when the composer is busy,
to avoid calls interleave. When the message is building, the composer
asks the HTML editor for the message content. This is done with a D-Bus
call to the WebKitWebProcess, which cannot block the main loop, because
the WebKit can do its own IPC calls there, thus the GUI would just block
in that case. Having the GUI updated during the wait for the message
content can also mean that the user will send the message in that time,
or the autosave will be triggered, where both can mean a use-after-free
of the variables being involved in the message building.

This had been reported downstream at:
https://bugzilla.redhat.com/show_bug.cgi?id=1538865