Try to preselect source type when creating new source
When creating new Address Book, Calendar, Memo List or Task List,
try to preselect the source type based on the currently selected
source in the active view, accounting also the clicked source
in the source selector.

This way, when creating a new calendar in the Calendar view the type
will match the one selected in the source selector, or when called
from the context menu of the source selector and it had been clicked
on the group name, then there will be preselected that group type,
if possible.