[ENameSelectorEntry] Plain text drag & drop between entries misbehaves
When drag&drop selected text between two name selector entries the first
drop could miss the text and lost it on focus and the consecutive
drops were not treated as separate addresses, which could break
the intention of the user. This change always adds dropped text as new
separate destinations.