Propagate trust prompt response within collection sources
Collection sources often access the same server, thus can have
the same SSL trust settings, which are inherited from the collection
source to the child sources when the child source is created for
the first time. There is no need to ask for a trust prompt with
every single child source when it accesses the same server, thus
this change propagates SSL trust into the collection source and
its children after user responses to it.