Unnecessary message parse when viewing message source
This could be noticed for example when viewing message source
of a PGP encrypted message, where a prompt for the key password
could be shown, even this message content wasn't used. It could
be used when having the message source opened and opening
the message in the message preview (EMailPartList is reused
in that case).