Composer autosave can stop working
In some cases, like when the autosave timer elapsed just it time
of the composer being "soft busy", or when there was a problem
with the editor itself, the autosave code left marked itself as
"waiting to save the changes", which was not true. In case of
the "soft busy" state it meant no further autosaving.