[Flatpak manifests] Build geocode-glib directly with meson
As geocode-glib is provided with a ./configure script the build system
in Flatpak had been chosen to be autotool, even the script calls meson
in the background. This script requires 'mesontest' binary, which had
been removed in meson 0.48.0. Using meson build system directly makes
the build work regardless of used meson version.