Clamp GSettings values before restoring window size
The GSettings values can be changed by anything, which means the user
can override the last stored window size to some large enough that
the gdk (or WebKitGTK+) will fail to allocated such large pixmap,
which can lead to a crash. This size clamp helps to avoid it.

Related to downstream bug:
https://bugzilla.redhat.com/show_bug.cgi?id=1657361