[Composer] Draft email can lose attachments when opened
Parts with Content-ID or Content-Location had been considered as inline
content and not added into real attachments, even when not referenced
in the HTML body. This addresses it in a way that it's only for images
and the worse case is to add it to both inline images and attachments,
to not lose it.

Reported downstream as:
https://bugzilla.redhat.com/show_bug.cgi?id=1666572