The 'Show' combo box hidden for small screens in the search bar ][
The previous change missed a fact that the EActionComboBox is used on
various places where always-enabled ellipsizing can cause issues, like
in the composer, where the combox of the message format and the paragraph
mode always used minimum size, thus made those unreadable. This change
enables ellipsizing in the EActionComboBox on demand, with the default
to not ellipsize, which keeps the behavior as before the change for
all but the search bar.