Ref/unref imbalance on a CamelFolder instance
When doing searches in the Mail view in other than Current Folder context
and then clearing the search text, the search folder in the background was
not freed, instead of it the real folder, into which the view had been
switched, had been unreffed one more time. That could cause some issue
in the future or at the quit of the application.

Few other things found during this fixed as well.

Related to a downstream bug report:
https://bugzilla.redhat.com/show_bug.cgi?id=1678440