Use GWeakRef when watching D-Bus bus names
With certain operations interleaving the callback to name-appeared
signal could be called after the object in the user_data had been freed.
Even when the free of the object unwatches the D-Bus name, it's too
late, because the signal callback is already scheduled in the main loop.

Reported downstream at:
https://bugzilla.redhat.com/show_bug.cgi?id=1677846