Fix a crash when filling mail threaded view
There could happen a crash when a folder received an update and it
had been shown in the message list in the threaded view. With a proper
thread interleaving the main thread could use the CamelFolderThread
structure, while another thread had been freeing its internal members.
The fix is to not re-use existing CamelFolderThread structure and always
build a new one. This makes the filling a bit quicker too.

Reported downstream at:
https://bugzilla.redhat.com/show_bug.cgi?id=1709367