Webview font size fixes
MR !14

* Allow floating point value for webview body font size (in pt). This
allow to respect the user defined UI font size in the webview part of
evolution. Uses an intermediate buffer with g_ascii_dtostr conversion to
avoid locale-related shortcomings of '%f'.

* Keep the default webview settings for base font size: this is mainly
used to compute sizes in em/rem, and the default values are standards
(16px for default-font-size and 13px for default-monospace-font-size).
Setting the base font size to the user defined size for UI make little
sense to me. Moreover, this should better play with HTML content where
fonts are sized exclusively using 'em' units (which is the recommended
way nowadays).

* Remove dead code after the previous change.