emoticon: Grab keyboard on the popup window
Grabbing the keyboard on the toplevel (regular) window will trigger the
shortcut inhibition on Wayland, which is not what we want.

But gtk is smart enough to bypass the shortcut inhibition when dealign
with popup windows, which is the case for the emoticon window, so we'd
better use that window for grabbing the keyboard.

Fixes: https://gitlab.gnome.org/GNOME/evolution/issues/485
Signed-off-by: Olivier Fourdan <ofourdan@redhat.com>