Remember source message information when re-editing in Drafts or Outbox
When a user saved a reply/forward message into Drafts and returned back
to it later the composed message lost an information about the source
message, thus when finally sent, the source message wasn't marked as
replied to or forwarded. Similarly when re-editing in the Outbox folder.
This is fixed by preserving related headers in the composer when
opening an existing message in either of the two folders.