evolution-csv-importer.c:getNextCSVEntry() memory leak
When EOF was reached, the function has not released line.

When (!parseLine (gci, contact, buf)) the function has not released str.

This changeset abolishes the variable GString *str.  The code essentially
moved the content from (GString*)line to str, and destroyed variable
'line'.  Now line is just used further in place of str.