M!33 - Update usage of stock_task-assigned-to icon
The Mail View Todo-Pane used on right mouse click to the action “New
Assigned Task” the icon (orange arrow), that is used to show that
somebody else created and assigned the task to the current user.

The icon here is changed to blue arrow (current user is organizer).

In the Event Window→Tasks bar and in the Task window, on right click,
show in the menu before “Assign Task” an appropriate icon (blue arrow).

Closes https://gitlab.gnome.org/GNOME/evolution/merge_requests/33