Fix few memory leaks in internal test programs
Connecting gtk_main_quit() directly to "delete-event" may not destroy
the window, thus make sure the callback returns FALSE, to have the window
properly destroyed.

In the test-source-selector specifically, when the opened client has
an active view, the view holds another reference to the client, thus
unreffing only the client in the hash table doesn't free it, neither
the view, which is assigned on the client instance, thus add a special
destroy callback and unset the view on the client first, then unref
the client.