calendar_config_init: Avoid runtime warning when called out of Evolution
The idea of freeing global memory on an EShell dispose works only if
the function is called in Evolution itself, where such EShell instance
exists, but it fails everywhere else, like in tests and also on the
evolution-data-server factory side, when used by evolution-ews.
The change doesn't really fix this, it either is able to free
the global memory as before, or the memory is leaked.