Fix a crash when changing Archive folder
This happened when for example a new folder had been created and after
it the archive folder had been changed in Preferences. The reason was
that there was left a signal handler for an EMFolderTreeModel, which had
been already freed. Disconnecting the signal handler in the dispose
of the EMFolderTreeModel fixes the crash.