EMFolderTreeModel: Let load folder structure on demand
It worked for newly created folders, but not after start, when a recursive,
aka complete, folder structure is requested from the CamelStore. After this
change the user can expand a folder node to load its subfolders also for
cases when the CamelStore sends a CamelFolderInfo with no child, but with
a CAMEL_FOLDER_CHILDREN flag, which notifies the caller that children
folders are available.

Related to https://gitlab.gnome.org/GNOME/evolution-ews/issues/61